refactor: use IETF language tags
Use IEFT tags (BCP 47) to indicate the language of a Unitex linguistic
resource. This is in accordance with the predominant use of these
language tags in modern computing standards.

Note: One of the best practices when creating language tags is to keep the tag
as short as possible. See http://www.langtag.net/tag-wisely.html for
more information on this.

IEFT tags related info:

- https://en.wikipedia.org/wiki/IETF_language_tag
- http://www.rfc-editor.org/rfc/bcp/bcp47.txt
- http://unicode.org/cldr/utility/languageid.jsp
- https://www.w3.org/International/articles/language-tags