Portug brazil - Remove obsolete entries from DELAF (#11)
* Dictionary of Portuguese without obsolete spelling

Uploading the dictionary of Portuguese (Brazil) updated for the spelling reform of 2009, without the entries in obsolete spelling. Cf. issue #9

* Delete DELAF_PB_2015.bin

DELAF_PB_2015.bin is replaced by DELAF_PB_2018.bin

* Delete DELAF_PB_2015.inf 

DELAF_PB_2015.inf is replaced by DELAF_PB_2018.inf

* Adding Portuguese DELAS without obsolete entries

Adding a version of the Portuguese (BR) DELAS updated for the spelling reform of 2009 and without obsolete entries

* Deleting Portuguese DELAS with obsolete entries

Deleting the version of the Portuguese (BR) DELAS updated for the spelling reform of 2009 but with obsolete entries

* Adding dictionary information file for DELAF_2018

Adding the dictionary information file for the DELAF_2018 dictionary

* Add dictionary information file for abbrev. dictionary

Adding the dictionary information file for the dictionary of abbreviations and acronyms

* Add dictionary information file for old DELAF

Adding a dictionary information file for the DELAF dictionary with the spelling of before the reform of 2009

* update dictionary information file for DELAF 2018

updating the dictionary information file for the DELAF_PB_2018 dictionary

* Updating system_dic.def

Updating system_dic.def for Portuguese (BR) since name of main DELAF changed